+++ swing/common-ui/src/main/java/is/codion/swing/common/ui/KeyEvents.java
@@ -113,20 +113,20 @@ public final class KeyEvents {
     Builder action(Action action);
 
     /**
-     * Builds the key event and enables it on the given component
-     * @param component the component
+     * Builds the key event and enables it on the given components
+     * @param components the components
      * @return this builder instance
      * @throws IllegalStateException in case no action has been set
      */
-    Builder enable(JComponent component);
+    Builder enable(JComponent... components);
 
     /**
-     * Disables this key event on the given component
-     * @param component the component
+     * Disables this key event on the given components
+     * @param components the components
      * @return this builder instance
      * @throws IllegalStateException in case no action has been set
      */
-    Builder disable(JComponent component);
+    Builder disable(JComponent... components);
   }
 
   private static final class DefaultBuilder implements Builder {
@@ -186,13 +186,21 @@ public final class KeyEvents {
     }
 
     @Override
-    public Builder enable(JComponent component) {
-      return enable(requireNonNull(component), actionMapKey(component));
+    public Builder enable(JComponent... components) {
+      for (JComponent component : requireNonNull(components)) {
+        enable(requireNonNull(component), actionMapKey(component));
+      }
+
+      return this;
     }
 
     @Override
-    public Builder disable(JComponent component) {
-      return disable(requireNonNull(component), actionMapKey(component));
+    public Builder disable(JComponent... components) {
+      for (JComponent component : requireNonNull(components)) {
+        disable(requireNonNull(component), actionMapKey(component));
+      }
+
+      return this;
     }
 
     private Object actionMapKey(JComponent component) {
@@ -214,24 +222,20 @@ public final class KeyEvents {
               .toString();
     }
 
-    private Builder enable(JComponent component, Object actionMapKey) {
+    private void enable(JComponent component, Object actionMapKey) {
       component.getActionMap().put(actionMapKey, action);
       component.getInputMap(condition).put(keyStroke, actionMapKey);
       if (component instanceof JComboBox<?>) {
         enable((JComponent) ((JComboBox<?>) component).getEditor().getEditorComponent(), actionMapKey);
       }
-
-      return this;
     }
 
-    private Builder disable(JComponent component, Object actionMapKey) {
+    private void disable(JComponent component, Object actionMapKey) {
       component.getActionMap().put(actionMapKey, null);
       component.getInputMap(condition).put(keyStroke, null);
       if (component instanceof JComboBox<?>) {
         disable((JComponent) ((JComboBox<?>) component).getEditor().getEditorComponent(), actionMapKey);
       }
-
-      return this;
     }
   }
 }